feat(homepage): hide championship data
Because data is now on Leverade and we'll never show it on this website.
It will be retrieved from Leverade to show on the new website.

This could be done more efficiently by completely removing the code
from the homepage instead of just hiding the rendering.
But it's not worth the effort as this website will soon be deprecated.